Исправление визуального бага, когда порядок рендера спрайта строений постоянно менялся
- Depth Entity теперь является float
- Добавление небольшого смещения к Depth спрайта строения

+++ palmesneo_village/palmesneo_village/Entities/Entity.cs
@@ -43,8 +43,8 @@ namespace palmesneo_village
             }
         }
 
-        private int depth = 0;
-        public int Depth
+        private float depth = 0;
+        public float Depth
         {
             get => depth;
             set
